+++ Polly/src/actions/polls.js
